fix: four macOS-only defects CUDA could never see, and the logit harness that could not run
Every fix here is on a path CUDA does not execute, which is why all four
survived a suite that is green on the CUDA box. Mac suite 263 passed/9 failed
-> 277 passed/0 failed.

logit_fidelity_mlx.py could not answer its own question. Two defects:

  * MLXDKVWrapper(model_id=...) omits `config`, a required positional arg, so
    every arm died before loading the model. Nothing had ever run.
  * the `dense` arm was DKV_COMPRESSED_DECODE=0, which on MLX forces exact
    full-KV attention at DECODE ONLY. The prompt is still read through
    block-sparse PREFILL (gated on manager._sparse_prefill and context length
    alone, mlx_dkv_wrapper.py:5493). This harness measures the FIRST decode
    step, a pure function of what prefill produced, so that arm compared
    DKV-prefill against DKV-prefill and would have reported a reassuringly
    small KL for a reason unrelated to fidelity. The control is now plain
    mlx_lm with DKV never loaded, matching mlx_needle_parity.py:104.

Added `blocks` and `max|dlogit|` columns because the headline number is
otherwise ambiguous: a KL of 0 from an empty pool means nothing, and
bit-identical logits would mean the compressed path never ran.

Result, Qwen2.5-1.5B @8k, 5 depths: MLX DKV baseline agrees with dense 5/5,
KL 5.135e-12, dense-top1 at rank 0.00, 168 compressed blocks, max|dlogit|
3.125e-02 = exactly one fp16 ULP. CUDA's baseline on the same prompt is 0/5
and KL 10.579. By MLX_TODO.md section 1's own pre-decided reading that makes
CUDA's gap a CUDA defect, not the price of compression.

native_triton_sparse_attn_decode returned a ZERO-WIDTH tensor for a zero-block
step. The N==0 dense-only fix of 2026-08-17 went into the Triton branch, but
the function short-circuits at `if not HAS_TRITON:` and returns before ever
reaching it -- so CUDA always skipped the broken branch and Apple silicon and
CPU always took it. Same _dense_only_attend guard now applies there.

decode_attention_metal rejects Long slot_indices (the guard added with
39a4a9d1, the fp16-RoPE-table fix). That commit converted anchor_indices at
the call site and left block_indices raw at all four sites, so the conversion
its own error message prescribes was never applied. _DKV_HAS_METAL_ATTN is
True on this machine, i.e. the branch is live here. Converted at all four,
plus the stale test that predates the guard.

MPS dense window was rotated TWICE, and not only in the validator. The
production _is_mps_decode path hands dense_k_assembled to the shader with
cos/sin, but those rows come from the blocks' active_k and are already
post-RoPE under DKV_ROTATED_POOL=1. `low` is the only preset keeping
rotated_pool=True (config.py:143) and also sets approximate_attn=True on
macOS (config.py:37) -- exactly the _is_mps_decode gate -- so `low` on Apple
silicon double-rotated its whole dense window. Both sites now consult
_pool_rotated_k(); empty cos/sin is the correct disable because
metal_runtime.mm:453 derives has_dense_rope separately from has_dense, so the
window is still attended, just not rotated. RoPE is orthogonal, so this
preserves every norm and corrupts only angles -- it cannot be caught by a
norm assertion, and test_mps_dense_rope_guard.py demonstrates exactly that
rather than merely describing it.

conftest now restores DKV_ROTATED_POOL and DKV_SVD_ENERGY per test.
config.py:814,819 EXPORT preset-derived values into the process environment
with setdefault(), and MLXKVBlockManager reads DKV_ROTATED_POOL at
construction (mlx_dkv_wrapper.py:1783). setdefault means the first config
built in the process wins for the whole process, so one test building a
config on a preset with rotated_pool=False silently flipped the pool frame
for every MLX manager afterwards -- test_residual_budget_clamp passed alone
and failed in the suite for that reason. Same isolation the
DKV_POOL_BUDGET_GB pin already does.

mlx_dkv_wrapper.py is NOT touched (HANDOFF section 0 rule 2).

+++ ACTIVE_RUNTIME/tests/test_sparse_residual.py
@@ -328,7 +328,13 @@ def test_metal_residual_and_fact_parity():
     )
     
     Q = torch.randn(num_heads, head_dim, device=device, dtype=torch.float16)
-    block_indices = torch.tensor([pool_idx], device=device, dtype=torch.long)
+    # int32, not long: this test drives dkv_core.decode_attention_metal DIRECTLY,
+    # and the shader reads slot_indices through a typed pointer. The binding grew
+    # an explicit dtype guard with the fp16-RoPE-table fix (39a4a9d1) and this
+    # test was not updated with it, so it has failed on every Mac since -- and
+    # CUDA never runs it, which is why it stayed unnoticed. The sibling
+    # test_decode_attention_metal_isolated.py already passes int32 here.
+    block_indices = torch.tensor([pool_idx], device=device, dtype=torch.int32)
     blk_sizes = torch.tensor([seq_len], device=device, dtype=torch.int32)
     anchor_indices = torch.tensor([0], device=device, dtype=torch.long)
